requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==2.4.0
django-modeltranslation==0.17.3
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==2.10
inflection==0.5.1
kombu==5.1.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.19
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
redis-server==6.0.9
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social==0.1.dev3
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
